+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Added the `array_sharded_ext::{ArrayShardedExt,ArrayShardedReadableExt}` array extension traits to simplify working with sharded arrays
+   - Abstracts the chunk grid to an "inner chunk grid" to simplify inner chunk retrieval
+   - Shard indexes are cached in a `ArrayShardedReadableExtCache`
+
 ### Changed
  - Allow float fill values to be created from int fill value metadata
  - Make `chunk_grid::{regular,rectangular}` public